separated joy header, applied clang formatting
