+++ launch/teleop_joy.launch
@@ -1,0 +1,9 @@
+<?xml version="1.0" ?>
+<launch>
+  <rosparam ns="robotont" file="$(find robotont_driver)/config/robotont_params.yaml" command="load" />
+  <node ns="robotont" pkg="joy" name="joy_node" type="joy_node">
+    <param name="deadzone" value="0.25" />
+    <param name="autorepeat_rate" value="5" />
+  </node>
+  <node ns="robotont" pkg="robotont_demos" name="teleop_joy" type="teleop_joy" output="screen" />
+</launch>
